% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Simulation many binomial experiments.
Perform a simulation of "iter" number of binomial experiments with n trials
per experiment where p is the probability of success in each trial.}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{the number of experiments to run}

\item{n}{the number of trials per experiment}

\item{p}{the probability of a success in a trial}
}
\value{
a table of the relative frequencies for each number of successes possible (given n)
}
\description{
Simulation many binomial experiments.
Perform a simulation of "iter" number of binomial experiments with n trials
per experiment where p is the probability of success in each trial.
}
\examples{
mybin(iter = 10000, n = 10, p = 0.35)
}
