% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{Bootstrap simulation technique}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{iter}{the number of iterations to run}

\item{x}{the original sample to resample}

\item{fun}{the statistic to generate using bootstrap}

\item{alpha}{the desired confidence level}

\item{...}{extra arguments}
}
\value{
a named list containing ci, the confidence interval, pte, the point estimate, fun, the function of the statistic to estimate, and the original sample x.
}
\description{
Bootstrap simulation technique
}
\examples{
myboot(iter = 10000, x = c(1), fun = "mean", alpha = 0.05)
}
