% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Demonstrate the Central Limit Theorem applied to the sampling distribution of the mean of the Poisson distribution.}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{the number of samples per sampling mean}

\item{iter}{the number of sampling means to calculate}

\item{lambda}{the lambda parameter for the lambda distribution in question}

\item{...}{additional parameters for the primary histogram plot}
}
\value{
NULL, nothing, zilch.
}
\description{
Demonstrate the Central Limit Theorem applied to the sampling distribution of the mean of the Poisson distribution.
}
\examples{
mycltp(n = 10, iter = 10000, lambda = 4)
}
