% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{Maximum likelihood estimate for normal distributions}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{vector of data for which to calculate the maximum likelihood estiamtes of mu and sigma}

\item{mu}{vector of values of mu to evaluate}

\item{sig}{vector of values of sigma to evaluate}

\item{...}{additional graphing parameters}
}
\value{
a named list containing x, the input vector of data, coords, the coordinates in the mu-sigma matrix for the mu and sigma estimates, and maxl, the maximum likelihood
}
\description{
Maximum likelihood estimate for normal distributions
}
\examples{
mymlnorm(x=c(10,12,13), mu=seq(10,14,l=1000), sig=seq(0.1,4,l=1000))
}
