% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Generates a plot of a normal curve with a specified µ and σ. Fills the area
under the curve up to 'a', and outputs the probability P(Y <= a) which is the
true area under the curve from -inf to 'a'.}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean of the normal distribution}

\item{sigma}{the standard deviation of the normal distribution}

\item{a}{the upper limit of the probability calculation (lower limit is always -inf)}
}
\value{
the numerical value of the probability (the area under the curve)
}
\description{
Generates a plot of a normal curve with a specified µ and σ. Fills the area
under the curve up to 'a', and outputs the probability P(Y <= a) which is the
true area under the curve from -inf to 'a'.
}
\examples{
myncurve(mu = 5, sigma = 2, a = 3)
}
