% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Optimum ticket sales models for airline tickets. Determines the optimal value for ticket sales when customers don't arrive based on gamma, the likelihood of overbooking.}
\usage{
ntickets(N = 200, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{the number of seats available on the plane.}

\item{gamma}{the likelihood of overbooking}

\item{p}{the probability that a customer arrives to board the plane}
}
\value{
a named list containing: nd the number of tickets to sell under the discrete model, nc the number of tickets to sell under the continuous model, N the number of seats available, p the probability a customer arrives to board, and gamma the likelihood of overbooking.
}
\description{
Optimum ticket sales models for airline tickets. Determines the optimal value for ticket sales when customers don't arrive based on gamma, the likelihood of overbooking.
}
\examples{
ntickets(N = 400, gamma = 0.02, p = 0.95)
}
