% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torreFun.R
\name{torreFun}
\alias{torreFun}
\title{Square a vector}
\usage{
torreFun(x)
}
\arguments{
\item{x}{a vector of quantitative values}
}
\value{
a vector of the values in x squared
}
\description{
Square a vector
}
\examples{
torreFun(x = 1:4)
}
